pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "video-analyzer"
version = "1.1.0"
description = "Offline video analysis using Ollama models and local Whisper"
authors = [{name = "Rohit Yelukati Mahendra", email = "mahendrarohittigon@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy>=2.0.2",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "requests>=2.32.3",
    "torch>=2.5.1",
    "transformers>=4.46.1",
    "librosa>=0.10.2.post1",
    "ffmpeg-python>=0.2.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0"
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
video-analyzer = "video_analyzer.cli:main"
video-analyzer-api = "video_analyzer.api_server:run"

[tool.hatch.build.targets.wheel]
packages = ["video_analyzer"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[project.urls]
Homepage = "https://github.com/ymrohit/openscenesense-ollama"
Repository = "https://github.com/ymrohit/openscenesense-ollama.git"
